# Conflicts:
#	Sky/Library/CurrentLayout.dwlt
#	Sky/Library/assetDatabase3
#	Sky/Library/metadata/00/00000000000000004000000000000000
#	Sky/Library/metadata/00/00000000000000006100000000000000
#	Sky/Library/metadata/38/38185d0338a894e609a39de524e52473
#	Sky/Library/metadata/f2/f25540c9e9e5847e0b4c18ca03ff0d2a

+++ Sky/Assets/Scripts/Memory/EndlessScore.cs
@@ -6,9 +6,9 @@ using System;
 public class EndlessScore : IComparable<EndlessScore> {
 
 	int score;          public int Score {get { return score; } }
-	decimal duration; public decimal Duration {get { return duration; } }
+	float duration; public float Duration {get { return duration; } }
 
-	public EndlessScore (int score, decimal duration){
+	public EndlessScore (int score, float duration){
 		this.score = score;
 		this.duration = duration;
 	}
